Merge Progress #7
You can view at line.notify branch for marge progress too. Some codes might not be the same between the twos. It is in decision process.

Merge Progress #5 and #6 was done in line.notify branch and previously are bugs fix and quality controlled.
This commit is merge process between line.notify branch and v0.2 branch.

    FIX:
      Realignment some codes to look better.
      Change login condition to old ways. (Using numbers instead of using distinct variables) For some files that haven't been affected yet.
      Replace $_SESSION['UserID'] with $_SESSION['master_id']
      Replace $_SESSION['User'] with $_SESSION['name']
      Change some outputs according to line.notify branch. (For some codes that haven't been affected yet.)
      Update ./admin/slidebar_admin.php according to newer commit.
      In ./admin/slidebar_admin.php remove declaration redundancy. (Already declare $_SESSION['name'] in ./login_db.php)
      Correct codes from typos in some files.